src/requirements.txt
alabaster==0.7.13
anndata==0.9.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer==2.1.1
click==8.1.3
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm==0.1.3
contourpy==1.0.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dgl-cuda11.3==0.9.1
docutils==0.19
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.5.9
fonttools==4.39.3
fqdn==1.5.1
glob2==0.7
h5py==3.8.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
igraph==0.10.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
leidenalg==0.9.1
libclang==16.0.0
littleutils==0.2.2
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
natsort==8.3.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
numpydoc==1.5.0
oauthlib==3.2.2
ogb==1.3.6
outdated==0.2.2
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
plotly==5.14.1
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
Pygments==2.15.1
pynndescent==0.5.10
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
rdkit-pypi==2022.9.5
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpy2==3.5.11
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scanpy==1.9.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
session-info==1.0.0
shapely==2.0.1
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
statsmodels==0.14.0
stdlib-list==0.8.0
sympy==1.11.1
tenacity==8.2.2
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.3.1
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
umap-learn==0.5.3
uri-template==1.2.0
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.14.1
zipp==3.15.0